add an option to load the config without validating it (using another yoshi-config/loadConfig) (#1028)
